requirements.txt
Flask
psycopg2-binary
python-dotenv
websockets
paho-mqtt